refactor: extract shared number literal parser
