+++ src/lexer/DecimalLiteralReader.js
@@ -1,4 +1,4 @@
-import { readDigits, isDigit } from './utils.js';
+import { readNumberLiteral, isDigit } from './utils.js';
 
 export function DecimalLiteralReader(stream, factory) {
   const startPos = stream.getPosition();
@@ -13,43 +13,26 @@ export function DecimalLiteralReader(stream, factory) {
     let value = '0' + stream.peek();
     stream.advance(); // 0
     stream.advance(); // d or D
-    value += readDigits(stream);
-    ch = stream.current();
-    if (ch === '.') {
-      value += '.';
-      stream.advance();
-      const digits = readDigits(stream);
-      if (digits.length === 0) {
-        stream.setPosition(startPos);
-        return null;
-      }
-      value += digits;
-      ch = stream.current();
-    }
+    const result = readNumberLiteral(stream, startPos, true);
+    if (!result) return null;
+    value += result.value;
     const endPos = stream.getPosition();
     return factory('DECIMAL', value, startPos, endPos);
   }
 
   // suffix form 123.45m or 123m
   if (isDigit(ch)) {
-    let value = readDigits(stream);
-    ch = stream.current();
-    if (ch === '.') {
-      value += '.';
-      stream.advance();
-      const digits = readDigits(stream);
-      if (digits.length === 0) {
-        stream.setPosition(startPos);
-        return null;
+    const result = readNumberLiteral(stream, startPos, true);
+    if (result) {
+      let { value, ch: next } = result;
+      if (next === 'm' || next === 'M') {
+        value += next;
+        stream.advance();
+        const endPos = stream.getPosition();
+        return factory('DECIMAL', value, startPos, endPos);
       }
-      value += digits;
-      ch = stream.current();
-    }
-    if (ch === 'm' || ch === 'M') {
-      value += ch;
-      stream.advance();
-      const endPos = stream.getPosition();
-      return factory('DECIMAL', value, startPos, endPos);
+    } else {
+      return null;
     }
   }
